requirements.txt
aiohttp==3.5.4
aiojobs==0.2.2
ansimarkup==1.4.0
app-state==0.3
async-timeout==3.0.1
attrs==19.1.0
better-exceptions-fork==0.2.1.post6
chardet==3.0.4
Click==7.0
colorama==0.4.1
django-environ==0.4.5
fastapi==0.12.1
getinstance==0.1
h11==0.8.1
httptools==0.0.13
idna==2.8
lockorator==0.1
loguru==0.2.5
multidict==4.5.2
psutil==5.6.1
pydantic==0.21
Pygments==2.3.1
sniffio==1.0.0
starlette==0.11.1
uvicorn==0.6.1
uvloop==0.12.2
websockets==7.0
yarl==1.3.0
